+++ libstorage/RocksDBStorage.cpp
@@ -46,7 +46,7 @@ using namespace dev::storage;
 using namespace rocksdb;
 
 Entries::Ptr RocksDBStorage::select(
-    h256, int, TableInfo::Ptr tableInfo, const string& key, Condition::Ptr condition)
+    h256, int64_t, TableInfo::Ptr tableInfo, const string& key, Condition::Ptr condition)
 {
     try
     {
